fix(odn): 取号 SQL 序号参数显式 ::int 定型,修裸参数被推断为 text 后恒取 001
102 部署首验暴露二段缺陷:SUBSTRING(code FROM $2 FOR $3) 的裸参数被 PG 按重载
解析成 POSIX 正则变体 substring(text,text,text),序号切片恒 NULL、MAX 恒 0,端点
对任何网格都返回 001(P91 存量 011/012 视而不见);本地 pgx v5.7.5 则在编码层
直接拒绝 int→text。两种驱动行为都不可接受,根因是参数类型留给 PG 推断。

改法:$2::int/$3::int 显式定型,强制解析到 substring(text,int,int);
102 真实 PG 上 pgx 双路径(PREPARE/EXECUTE 与 pgx 客户端)复核 P91 max=12。

补 TestNextFacilityCode_Integration(BOSS_PG_TEST_DSN 门控,网格 96/97 隔离,
不碰存量 91/98 数据):真实 pgxpool 走完整取号路径,守住 pgxmock 够不到的
语法/参数定型层;自清不依赖 t.Cleanup——它跑在 pool.Close 之后静默失效,
存量集成测试正是因此给 102 泄漏了 P98xxx 造数(顺带发现,另行跟进)。

+++ internal/domain/odn/nextcode.go
@@ -67,8 +67,10 @@ func nextFacilityCode(ctx context.Context, q seqRowQuerier, kind string, gridCod
 		return "", err
 	}
 	var maxSeq int
+	// $2/$3 必须 ::int 显式定型:裸参数会被 PG 推断成 text,解析到 POSIX 正则变体
+	// substring(text,text,text),切片结果恒 NULL,取号永远返回 001(102 实测踩坑)。
 	if err := q.QueryRow(ctx,
-		`SELECT COALESCE(MAX(SUBSTRING(code FROM $2 FOR $3)::int),0)
+		`SELECT COALESCE(MAX(SUBSTRING(code FROM $2::int FOR $3::int)::int),0)
 			FROM odn_facility WHERE code ~ $1`,
 		spec.pattern, spec.offset, spec.digits).Scan(&maxSeq); err != nil {
 		return "", fmt.Errorf("odn: next facility code %s: %w", kind, err)
